requirements.txt
# Apache Airflow
apache-airflow==2.8.1

# Snowflake provider for Airflow
apache-airflow-providers-snowflake==5.2.1

# Snowflake connector
snowflake-connector-python==3.7.0

# Additional dependencies
cryptography==42.0.0
